requirements.txt
alembic==1.9.3
blinker==1.5
boto3==1.26.68
botocore==1.29.68
click==8.1.3
dnspython==2.3.0
email-validator==1.3.1
Flask==2.2.2
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
greenlet==2.0.2
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
Mako==1.2.4
MarkupSafe==2.1.2
mysqlclient==2.1.1
PyJWT==2.6.0
python-dateutil==2.8.2
python-dotenv==0.21.1
s3transfer==0.6.0
six==1.16.0
SQLAlchemy==2.0.3
typing_extensions==4.4.0
urllib3==1.26.14
Werkzeug==2.2.2
WTForms==3.0.1
zipp==3.13.0
